+++ server/api/registre/inventar/[id].delete.ts
@@ -1,8 +1,17 @@
 import { RegistruInventar } from "../../../models/RegistruInventar";
+import { validateMongoId } from "../../../utils/validation";
 
 export default defineEventHandler(async (event) => {
   const user = event.context.user;
-  const id = getRouterParam(event, "id");
+
+  if (!user) {
+    throw createError({
+      statusCode: 401,
+      message: "Neautentificat",
+    });
+  }
+
+  const id = validateMongoId(event, "id");
 
   const entry = await RegistruInventar.findOneAndDelete({
     _id: id,
